Extract CleaningService
